requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
chromadb>=0.4.0
faiss-cpu>=1.7.0
pdfplumber>=0.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
markdown2>=2.4.0
python-docx>=1.1.0
streamlit>=1.28.0
# Optional: sentence-transformers>=2.2.0 (for better embeddings)
# Optional: torch>=1.9.0 (for sentence-transformers)
